solve issue 27 using a sentinel name that no one else will ever use
problem was that in, e.g. Fake(path, **attributes), one of the
attributes may also be called "path" - why not?

the solution is to rename the positional parameter path to
path_a62df12dd67848be82c505d63b928725, which is a uuid and
thus it will never occur again in our lifetime.